// src/lib/uploader/interface.ts
export interface IPluginConfig {
  name: string;
  type: string;
  required: boolean;
  default?: any;
  alias?: string;
  message?: string;
  prefix?: string; // for cli options
  [propName: string]: any;
}